+++ src/assets/index.js
@@ -1,9 +1,14 @@
 import bidenImg from './biden.jpeg';
 import clintonImg from './clinton.jpeg';
+import demImg from './demlogo.jpeg';
+import grnImg from './grnlogo.png';
+import hawkinsImg from './hawkins.jpeg';
+import indImg from './indlogo.jpeg';
 import johnsonImg from './johnson.jpg';
 import jorgensenImg from './jorgensen.png';
-import hawkinsImg from './hawkins.jpeg';
+import libImg from './liblogo.png';
 import mcmullinImg from './mcmullin.jpeg';
+import repImg from './goplogo.png';
 import steinImg from './stein.jpg';
 import trumpImg from './trump.jpeg';
 import westImg from './west.jpeg';
@@ -11,10 +16,15 @@ import westImg from './west.jpeg';
 export {
   bidenImg,
   clintonImg,
+  demImg,
+  grnImg,
+  hawkinsImg,
+  indImg,
   johnsonImg,
   jorgensenImg,
-  hawkinsImg,
+  libImg,
   mcmullinImg,
+  repImg,
   steinImg,
   trumpImg,
   westImg