+++ src/constants/candidatesData-2024.js
@@ -1,0 +1,22 @@
+export const CANDIDATES_DATA = [
+  {
+    "name": "Rep",
+    "party": "gop"
+  },
+  {
+    "name": "Dem",
+    "party": "dem" 
+  },
+  {
+    "name": "Lib",
+    "party": "lib"
+  },
+  {
+    "name": "Grn",
+    "party": "grn"
+  },
+  {
+    "name": "Ind",
+    "party": "ind"
+  }
+];