+++ src/constants/index.js
@@ -1,6 +1,6 @@
-import { CANDIDATES_DATA } from './candidatesData-2020';
+import { CANDIDATES_DATA } from './candidatesData-2024';
 import { PARTY_MAP } from './partyData';
-import { STATES_DATA } from './statesData';
+import { STATES_DATA } from './statesData-2020';
 
 export {
   CANDIDATES_DATA,